cursor: don't synthesize relative motion events from absolute events
When the cursor position is locked via constraint, the computed deltas
continue to get larger as the absolute event position diverges further
from the locked position. This led to the mouse pointer going crazy in
applications that use the relative events, such as games under
Wine/Wayland.

There can be other situations as well where an unexpectedly large delta
is computed. Since the purpose of synthesizing the relative events in
the first place isn't clear, let's just not.